auth_api/python/requirements.txt
Flask
mysql-connector-python
pyjwt[crypto]